gstreamer1.0-plugins-bad: fix compilation when wayland is enabled
When compiling with wayland enabled in PACKAGECONFIG, we hit the following
compile issue:

make[3]: *** No rule to make target 'viewporter-protocol.c', needed by 'all'.
Stop.

This issue was introduced in commit d164012c1, when we imported gst 1.10 into
meta-backports. The same issue was observed in master, and it was fixed by
adding a dependency on wayland-protocols and by setting the WAYLAND_SYSROOT_PATH
properly, however in master sysroot is now per-recipe , and the EXTRA_OECONF, in
morty, needed to be modified to used the previous sysroot variable.

Change-Id: I3019b4a9c50a9faf9f070d5589ea16932adaffc7
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>