packagegroup-rpb-{weston,x11}: add kmscube
kmscube is a convenient test app that we want in our images. however while
kmscube was merged in oe-core master, it is not there in oe-core morty
branch. So it was backported in meta-backports.

When using both meta-backports and meta-rpb, this commit will make sure that
kmscube is added into packagegroup-rpb-x11 and packagegroup-rpb-weston.

This commit is a 'backport' from meta-rpb:

ef5def8 packagegroup-rpb-{x11,weston}: add kmscube

Change-Id: I25bc9a6873580ed072fd022e50d553411beccb8f
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
(cherry picked from commit 373e2988dd54712231492a8e1344275c8e5345b1)