layer.conf: increase layer priority
meta-backport is used mostly to overlay recipes with newer version of
software. It makes sense for meta-backports to have a priority as high as
possible so that its recipes will *always* be prefered.

For example, meta-oe has a priority of 6, and having meta-backports at 5, then
all the recipes that we overlay in meta-backports are not being
considered. e.g. efivar, glmark2, ..

Change-Id: I70439ee7fd0476e2b508fa4a4eda4c0585631d1b
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>