ltp: bump to release 20180515, for real
The following patches have made it to upstream:
* 0001-configure-Fix-default-value-of-without-numa-switch-i.patch
* 0001-configure-add-knob-to-control-numa-support.patch
* 0003-Add-knob-to-control-tirpc-support.patch
* 0010-replace-__BEGIN_DECLS-and-__END_DECLS.patch
* 0040-safe_macros-make-is_fuse-return-zero-if-fs_type-is-N.patch

These patches have been refreshed in order to apply cleanly
and without any fuzz:
* 0004-build-Add-option-to-select-libc-implementation.patch
* 0008-Check-if-__GLIBC_PREREQ-is-defined-before-using-it.patch
* 0021-Define-_GNU_SOURCE-for-MREMAP_MAYMOVE-definition.patch
* 0028-rt_sigaction.h-Use-sighandler_t-instead-of-__sighand.patch
* 0036-testcases-network-nfsv4-acl-acl1.c-Security-fix-on-s.patch

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>

We ship the open posix test suite already but it isn't compiled.
Let's build it and make it ready to use on the device.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>

Add new patch from OE-core:
* 0040-read_all-Define-FNM_EXTMATCH-if-not-already-like-und.patch
and one that is yet to be merged upstream:
* 0041-cve-2017-5669-shmat-for-0-or-PAGESIZE-with-RND-flag-.patch

Add new RDEPENDS: procps.

Change a few things in the recipe, chiefly the part where it
deals with the packaging of a static library.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>

Include the following change:
  3084ff86 ltp: set -fomit-frame-pointer explicitly for x86-64
  Mingli Yu <Mingli.Yu@windriver.com>

Also sync the patch that Naresh sent (0041) as it was taken by
OE-core.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>

Change-Id: I6d34dd00f86fd0fe98bff52f3a8d2b7ced96361a
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>